Get info from a Vertiv Avocent 8000
Using the REST API on the Avocent, this script will grab the serial port number, the Alias SSH port, verify the status, list the pinout & speed of the serial line and grab the name of the device connected to the port. This is useful if you configure the Avocent to dynamically find the attached device. This is the base file to larger project.